Fixed rsync syncer.
Rsync syncer with SSH mode requires key pair and the public key
authorized in the destination server. The implementation then needs to
store the public key in a databag and force the destination server to
read the new key from the databag. Although this can be implemented, we
prefer to use rsync daemon on the destination server, which is now the
default.

This cookbook still supports SSH mode for rsync syncer, but the keys
must be set separately.