Refactored storages logic.
(This commit completes the previous one which has omited some files by
mistake).

In libraries/storages.rb file was added with the storage management
logic implemented as classes. As a consquence, storage section in
model template was simplified and new storages can now be easily added
using a custom DSL.